+++ main/src/main/java/net/oneandone/maven/summon/Maven.java
@@ -54,7 +54,7 @@ import java.util.Properties;
 
 public class Maven implements AutoCloseable {
     public static Maven create() throws IOException {
-        return new Maven(Config.create());
+        return new Config().build();
     }
 
     //--
@@ -75,16 +75,16 @@ public class Maven implements AutoCloseable {
     // the deprecated org.apache.maven.artifact.repository.ArtifactRepository class, so deprecation warnings are unavoidable.
     private final ProjectBuilder projectBuilder;
 
-    public Maven(Config config) {
-        this(config, LegacyRepositories.create(config));
+    public Maven(Repositories repositories) {
+        this(repositories, LegacyRepositories.create(repositories));
     }
 
-    public Maven(Config config, LegacyRepositories legacyRepositories) {
-        this.container = config.container();
-        this.repositorySystem = config.repositorySystem();
-        this.repositorySession = config.repositorySession();
-        this.remote = config.repositories();
-        this.projectBuilder = config.projectBuilder();
+    public Maven(Repositories repositories, LegacyRepositories legacyRepositories) {
+        this.container = repositories.container();
+        this.repositorySystem = repositories.repositorySystem();
+        this.repositorySession = repositories.repositorySession();
+        this.remote = repositories.repositories();
+        this.projectBuilder = repositories.projectBuilder();
         this.legacy = legacyRepositories;
     }